// src/app/models/sms/PutAvaliaçõesOcupacionaisFuncoesModel.ts
export class PutAvaliaçõesOcupacionaisFuncoesModel{
    idAvaliacoes_funcao: string = '';
    valor_obtido: string = '';
    limite_de_tolerancia: string = '';
	tempo_de_exposicao: string = '';
	metodo: string = '';
    conclusao_da_exposicao: string = '';
	equipamento_de_medicao: string = '';
	modelo: string = '';
    serie: string = '';
    calibracao: string = '';
    nome_avaliacoes_funcao: string = '';
    tipo_avaliacoes_funcao: string = '';
    descricao_avaliacoes_funcao: string = '';
     }